fix warning
git-svn-id: https://svn.salilab.org/imp/trunk@10439 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/algebra/include/VectorD.h
@@ -83,7 +83,7 @@ public:
 
   /** \throw ValueException if f.size() is not appropriate.*/
   VectorD(const Floats &f) {
-    if (D!=-1 && f.size() != D) {
+    if (D!=-1 && static_cast<int>(f.size()) != D) {
       IMP_THROW("Expected " << D << " but got " << f.size(),
                 ValueException);
     }
